updated navigation - updated the bottomTabs icons - created a @types folder for each navigation stack (contains the stackParamList, navigationProp) - created a @types folder for the bottomTabs (contains the bottomTabsParamList)
