+++ App.tsx
@@ -1,4 +1,4 @@
-import MainNavigator from "./src/navigation/Navigator"
+import MainNavigator from "./src/navigation/bottomTabs/BottomTabs"
 
 const App = (): JSX.Element => {
     return <MainNavigator />
