--- src/@types/patients/index.ts
@@ -1,8 +1,0 @@
-type Patient = {
-    patientId: number | string
-    name: string
-    age: number
-    image: string
-}
-
-export type { Patient }
