--- src/@types/products/index.ts
@@ -1,9 +1,0 @@
-type Product = {
-    productId: number
-    name: string
-    type: string
-    therapeuticClass: string
-    image: string
-}
-
-export type { Product }
