+++ src/navigation/bottomTabs/@types/index.ts
@@ -1,0 +1,9 @@
+/** ========================  BOTTOM TABS PARAMS LIST  ======================== */
+
+export type BottomTabsParamsList = {
+    Dashboard: undefined
+    Patients: undefined
+    Prescriptions: undefined
+    Products: undefined
+    Settings: undefined
+}
