--- src/navigation/options.ts
@@ -1,8 +1,0 @@
-import { StackNavigationOptions } from "@react-navigation/stack"
-
-const stackNavigatorOptions: StackNavigationOptions = {
-    headerShown: false,
-    animationEnabled: false,
-}
-
-export { stackNavigatorOptions }
