+++ src/navigation/stacks/patientStack/PatientsStack.tsx
@@ -1,15 +1,16 @@
 import { NavigationContainer } from "@react-navigation/native"
 import { createStackNavigator } from "@react-navigation/stack"
 import { PatientsList, PatientDetails } from "../../../screens"
-import { stackNavigatorOptions } from "../../options"
-import { PatientStackParamsList } from "../../../@types/navigation"
+import { PatientStackParamsList } from "./@types"
 
 const Stack = createStackNavigator<PatientStackParamsList>()
 
 const PatientsStack = () => {
     return (
         <NavigationContainer independent={true}>
-            <Stack.Navigator initialRouteName="PatientsList" screenOptions={stackNavigatorOptions}>
+            <Stack.Navigator
+                initialRouteName="PatientsList"
+                screenOptions={{ headerShown: false, animationEnabled: false }}>
                 <Stack.Screen name="PatientsList" component={PatientsList} />
                 <Stack.Screen name="PatientDetails" component={PatientDetails} />
             </Stack.Navigator>
