+++ src/navigation/stacks/prescriptionsStack/PrescriptionsStack.tsx
@@ -1,15 +1,16 @@
 import { NavigationContainer } from "@react-navigation/native"
 import { createStackNavigator } from "@react-navigation/stack"
 import { PrescriptionsList, PrescriptionDetails, AddPrescription } from "../../../screens"
-import { stackNavigatorOptions } from "../../options"
-import { PrescriptionsStackParamsList } from "../../../@types/navigation"
+import { PrescriptionsStackParamsList } from "./@types"
 
 const Stack = createStackNavigator<PrescriptionsStackParamsList>()
 
 const PrescriptionsStack = () => {
     return (
         <NavigationContainer independent={true}>
-            <Stack.Navigator initialRouteName="PrescriptionsList" screenOptions={stackNavigatorOptions}>
+            <Stack.Navigator
+                initialRouteName="PrescriptionsList"
+                screenOptions={{ headerShown: false, animationEnabled: false }}>
                 <Stack.Screen name="PrescriptionsList" component={PrescriptionsList} />
                 <Stack.Screen name="PrescriptionDetails" component={PrescriptionDetails} />
                 <Stack.Screen name="AddPrescription" component={AddPrescription} />
