+++ src/navigation/stacks/productsStack/ProductsStack.tsx
@@ -1,15 +1,16 @@
 import { NavigationContainer } from "@react-navigation/native"
 import { createStackNavigator } from "@react-navigation/stack"
 import { Products, ProductDetails } from "../../../screens"
-import { stackNavigatorOptions } from "../../options"
-import { ProductsStackParamsList } from "../../../@types/navigation"
+import { ProductsStackParamsList } from "./@types"
 
 const Stack = createStackNavigator<ProductsStackParamsList>()
 
 const ProductsStack = () => {
     return (
         <NavigationContainer independent={true}>
-            <Stack.Navigator initialRouteName="ProductsList" screenOptions={stackNavigatorOptions}>
+            <Stack.Navigator
+                initialRouteName="ProductsList"
+                screenOptions={{ headerShown: false, animationEnabled: false }}>
                 <Stack.Screen name="ProductsList" component={Products} />
                 <Stack.Screen name="ProductDetails" component={ProductDetails} />
             </Stack.Navigator>
