+++ src/navigation/stacks/settingsStack/SettingsStack.tsx
@@ -1,15 +1,16 @@
 import { NavigationContainer } from "@react-navigation/native"
 import { createStackNavigator } from "@react-navigation/stack"
 import { SettingsList, Profile, EditProfile, EditClinic } from "../../../screens"
-import { stackNavigatorOptions } from "../../options"
-import { SettingsStackParamsList } from "../../../@types/navigation"
+import { SettingsStackParamsList } from "./@types"
 
 const Stack = createStackNavigator<SettingsStackParamsList>()
 
 const SettingsStack = () => {
     return (
         <NavigationContainer independent={true}>
-            <Stack.Navigator initialRouteName="SettingsList" screenOptions={stackNavigatorOptions}>
+            <Stack.Navigator
+                initialRouteName="SettingsList"
+                screenOptions={{ headerShown: false, animationEnabled: false }}>
                 <Stack.Screen name="SettingsList" component={SettingsList} />
                 <Stack.Screen name="Profile" component={Profile} />
                 <Stack.Screen name="EditProfile" component={EditProfile} />
